refactor: move static codelist helpers into core BaseStaticResource
