+++ packages/core/resources/lang/de/fields.php
@@ -1,0 +1,7 @@
+<?php
+
+return [
+    'code' => 'Code',
+    'common_name' => 'Bezeichnung',
+    'description' => 'Beschreibung',
+];
