+++ packages/core/resources/lang/en/fields.php
@@ -1,0 +1,7 @@
+<?php
+
+return [
+    'code' => 'Code',
+    'common_name' => 'Common Name',
+    'description' => 'Description',
+];
